chore: extract price impact logic to hook
